chore: changed deployment
