+++ go.mod
@@ -3,22 +3,14 @@ module github.com/candidatos-info/site
 go 1.14
 
 require (
-	cloud.google.com/go v0.68.0 // indirect
-	cloud.google.com/go/bigquery v1.10.0 // indirect
-	cloud.google.com/go/datastore v1.3.0
-	cloud.google.com/go/pubsub v1.6.1 // indirect
-	cloud.google.com/go/storage v1.11.0 // indirect
-	github.com/aws/aws-sdk-go v1.35.4
+	github.com/aws/aws-sdk-go v1.35.4 // indirect
 	github.com/candidatos-info/descritor v0.0.0-20201006195312-a0b520a58186
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/gobeam/mongo-go-pagination v0.0.2
-	github.com/gocarina/gocsv v0.0.0-20200925213129-04be9ee2e1a2
 	github.com/golang/snappy v0.0.2 // indirect
 	github.com/klauspost/compress v1.11.1 // indirect
-	github.com/kr/pretty v0.2.1 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	github.com/labstack/echo v3.3.10+incompatible
-	github.com/labstack/echo/v4 v4.1.17
 	github.com/labstack/gommon v0.3.0 // indirect
 	github.com/mattn/go-colorable v0.1.8 // indirect
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
@@ -30,12 +22,7 @@ require (
 	golang.org/x/net v0.0.0-20201006153459-a7d1128ccaa0 // indirect
 	golang.org/x/sync v0.0.0-20200930132711-30421366ff76 // indirect
 	golang.org/x/sys v0.0.0-20201006155630-ac719f4daadf // indirect
-	golang.org/x/text v0.3.3
-	golang.org/x/tools v0.0.0-20201006174853-567bb5a4fa4e // indirect
-	google.golang.org/genproto v0.0.0-20201006033701-bcad7cf615f2 // indirect
 	gopkg.in/check.v1 v1.0.0-20200902074654-038fdea0a05b // indirect
-	gopkg.in/mgo.v2 v2.0.0-20190816093944-a6b53ec6cb22
 	gopkg.in/yaml.v2 v2.3.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776 // indirect
-	honnef.co/go/tools v0.0.1-2020.1.5 // indirect
 )
